Improving runtime editing of sensors and constraints

+++ automation/json/JsonStreamWriter.h
@@ -55,8 +55,8 @@ class JsonStreamWriter
   void clearChecksum() { checksum = 0; }
   unsigned long getChecksum() { return checksum; }
 
-  int depth;
-  int beginStringObjByteCnt;
+  unsigned int depth;
+  long beginStringObjByteCnt;
 
   JsonStreamWriter(OutputStreamPrinter& impl, int depth = 0) :
     impl(impl),
